cdk/requirements.txt
aws-cdk-lib==2.134.0
constructs>=10.0.0,<11.0.0
aws_cdk.aws_batch
aws_cdk.aws_ecs
aws_cdk.aws_ec2
aws_cdk.aws_cloudwatch